EID-918: Renaming directory config to configuration
Authors: @kgarwood, @felixmarcusmillne
