+++ src/config-test/java/uk/gov/ida/metadataaggregator/EnvironmentConfigTest.java
@@ -4,9 +4,9 @@ import org.junit.Before;
 import org.junit.Test;
 import org.opensaml.core.config.InitializationException;
 import org.opensaml.core.config.InitializationService;
-import uk.gov.ida.metadataaggregator.config.MetadataSourceConfiguration;
-import uk.gov.ida.metadataaggregator.config.ConfigSourceException;
-import uk.gov.ida.metadataaggregator.config.MetadataSourceConfigurationLoader;
+import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfiguration;
+import uk.gov.ida.metadataaggregator.configuration.ConfigSourceException;
+import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfigurationLoader;
 
 import java.net.URL;
 import java.util.Collection;
