+++ src/main/java/uk/gov/ida/metadataaggregator/MetadataAggregatorApplication.java
@@ -5,7 +5,7 @@ import com.squarespace.jersey2.guice.JerseyGuiceUtils;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
-import uk.gov.ida.metadataaggregator.config.MetadataAggregatorConfiguration;
+import uk.gov.ida.metadataaggregator.configuration.MetadataAggregatorConfiguration;
 
 public class MetadataAggregatorApplication extends Application<MetadataAggregatorConfiguration> {
