+++ src/test/java/uk/gov/ida/metadataaggregator/MetadataAggregatorTest.java
@@ -3,7 +3,7 @@ package uk.gov.ida.metadataaggregator;
 import org.junit.Before;
 import org.junit.Test;
 import org.opensaml.saml.saml2.metadata.EntityDescriptor;
-import uk.gov.ida.metadataaggregator.config.MetadataSourceConfiguration;
+import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfiguration;
 import uk.gov.ida.metadataaggregator.metadatasource.CountryMetadataResolver;
 import uk.gov.ida.metadataaggregator.metadatasource.MetadataSourceException;
 import uk.gov.ida.metadataaggregator.metadatastore.MetadataStore;
